refactore articleForm and use context
