+++ src/screens/ArticleForm/component/MDEditor/index.jsx
@@ -1,12 +1,15 @@
-import React, { useRef, useEffect, useState } from 'react';
+import React, { useRef, useContext } from 'react';
 import { useTheme } from '@material-ui/core';
 import ToolPanel from './ToolPanel';
 import useStyles from './styles';
+import { FormContext } from '@screens/ArticleForm/context/FormContext';
 
-const MDEditor = ({ handleMDText, mdText, setCursorPostion, setPanelValue, setRef }) => {
+const MDEditor = () => {
   const theme = useTheme();
   const classes = useStyles({ theme });
   const textareaRef = useRef();
+  const { setCursorPostion, setToolPanelValue, textareaValue, setTextareaValue, toolPanelValue } =
+    useContext(FormContext);
 
   function handleFocus(evt) {
     setCursorPostion(evt.target.selectionEnd);
@@ -19,10 +22,21 @@ const MDEditor = ({ handleMDText, mdText, setCursorPostion, setPanelValue, setRe
     evt.target.style.height = evt.target.scrollHeight + 'px';
   }
 
+  function handleChangeMD(evt) {
+    setCursorPostion(evt.target.selectionEnd);
+    evt.target.selectionEnd = evt.target.selectionEnd + toolPanelValue.length;
+    setTextareaValue(evt.target.value);
+    setToolPanelValue('');
+  }
+
+  function setRef(ref) {
+    ref.current?.focus();
+  }
+
   return (
     <div className={classes.root}>
       <ToolPanel
-        setPanelValue={setPanelValue}
+        setPanelValue={setToolPanelValue}
         setRef={setRef}
         textareaRef={textareaRef}
       ></ToolPanel>
@@ -31,8 +45,8 @@ const MDEditor = ({ handleMDText, mdText, setCursorPostion, setPanelValue, setRe
           className={classes.textArea}
           aria-label="Post Content"
           placeholder="Write your post content here..."
-          value={mdText}
-          onChange={handleMDText}
+          value={textareaValue}
+          onChange={handleChangeMD}
           onClick={handleFocus}
           onKeyDown={handleFocus}
           onKeyPress={handleChangeHeight}
