+++ src/screens/ArticleForm/component/Tags/types.d.ts
@@ -1,0 +1,4 @@
+export interface TagType {
+  name: string;
+  rules_html: DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> | string | null;
+}
